production models

+++ flaskr/nscope/models.py
@@ -33,7 +33,7 @@ class Sequence(db.Model):
 
     id = db.Column(db.String, unique=True, nullable=False, primary_key=True)
     name = db.Column(db.String, unique=False, nullable=True)
-    first_100_entries = db.Column(db.ARRAY(db.INTEGER), unique=False, nullable=False)
+    first_100_entries = db.Column(db.ARRAY(db.BigInteger), unique=False, nullable=False)
 
     @classmethod
     def get_seq_by_id(self, id):
